requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.3
dateparser==1.2.0
idna==3.6
lxml==5.1.0
pdfminer.six==20231228
pycparser==2.21
pycryptodome==3.20.0
pymystem3==0.2.0
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.2.1
